+++ test/test_granularity_convert/CMakeLists.txt
@@ -1,0 +1,4 @@
+target_sources(plateau_test PRIVATE
+        "${CMAKE_CURRENT_SOURCE_DIR}/test_granularity_converter.cpp"
+        "${CMAKE_CURRENT_SOURCE_DIR}/model_convert_test_patterns.cpp"
+)
