+++ test/test_granularity_convert/test_granularity_converter.h
@@ -1,6 +1,6 @@
 #pragma once
 #include "gtest/gtest.h"
-#include "test_granularity_converter_model_for_test.h"
+#include "model_convert_test_patterns.h"
 
 #include <plateau/granularity_convert/granularity_converter.h>
 #include <plateau/polygon_mesh/mesh_extractor.h>
@@ -12,8 +12,4 @@
 
 class GranularityConverterTest : public ::testing::Test {
 public:
-    ModelForTest createTestModelOfArea();
-    ModelForTest createTestModelOfAtomic();
-    ModelForTest createTestModelOfPrimary();
-    ModelForTest createTestModelOfArea_OnlyRoot();
 };
